+++ src/api/emit_support_code_messages_spec.ts
@@ -13,7 +13,10 @@ import { SupportCodeLibrary } from '../support_code_library_builder/types'
 import TestCaseHookDefinition from '../models/test_case_hook_definition'
 import TestRunHookDefinition from '../models/test_run_hook_definition'
 import { SourcedParameterTypeRegistry } from '../support_code_library_builder/sourced_parameter_type_registry'
-import { emitMetaMessage, emitSupportCodeMessages } from './helpers'
+import {
+  emitMetaMessage,
+  emitSupportCodeMessages,
+} from './emit_support_code_messages'
 
 const noopFunction = (): void => {
   // no code
@@ -54,7 +57,7 @@ function testEmitSupportCodeMessages(
   return envelopes
 }
 
-describe('helpers', () => {
+describe('emit_support_code_messages', () => {
   describe('emitMetaMessage', () => {
     it('emits a meta message', async () => {
       const envelopes: messages.Envelope[] = []
@@ -66,6 +69,7 @@ describe('helpers', () => {
       expect(envelopes[0].meta.implementation.name).to.eq('cucumber-js')
     })
   })
+
   describe('emitSupportCodeMessages', () => {
     it('emits messages for parameter types', () => {
       const parameterTypeRegistry = new SourcedParameterTypeRegistry()
@@ -81,7 +85,8 @@ describe('helpers', () => {
         {
           line: 4,
           uri: 'features/support/parameter-types.js',
-        }
+        },
+        0
       )
 
       const envelopes = testEmitSupportCodeMessages({
@@ -117,6 +122,7 @@ describe('helpers', () => {
             id: '0',
             line: 9,
             options: {},
+            order: 0,
             uri: 'features/support/cukes.js',
             keyword: 'Given',
             pattern: 'I have {int} cukes in my belly',
@@ -147,6 +153,7 @@ describe('helpers', () => {
       ]
       expect(envelopes).to.deep.eq(expectedEnvelopes)
     })
+
     it('emits messages for step definitions using regular expressions', () => {
       const envelopes = testEmitSupportCodeMessages({
         stepDefinitions: [
@@ -156,6 +163,7 @@ describe('helpers', () => {
             id: '0',
             line: 9,
             options: {},
+            order: 0,
             uri: 'features/support/cukes.js',
             keyword: 'Given',
             pattern: /I have (\d+) cukes in my belly/,
@@ -186,6 +194,7 @@ describe('helpers', () => {
       ]
       expect(envelopes).to.deep.eq(expectedEnvelopes)
     })
+
     it('emits messages for test case level hooks', () => {
       const envelopes = testEmitSupportCodeMessages({
         beforeTestCaseHookDefinitions: [
@@ -198,6 +207,7 @@ describe('helpers', () => {
               name: 'before hook',
               tags: '@hooks-tho',
             },
+            order: 0,
             uri: 'features/support/hooks.js',
           }),
         ],
@@ -210,6 +220,7 @@ describe('helpers', () => {
             options: {
               name: 'after hook',
             },
+            order: 1,
             uri: 'features/support/hooks.js',
           }),
           new TestCaseHookDefinition({
@@ -218,6 +229,7 @@ describe('helpers', () => {
             id: '2',
             line: 11,
             options: {},
+            order: 2,
             uri: 'features/support/hooks.js',
           }),
         ],
@@ -269,6 +281,7 @@ describe('helpers', () => {
       ]
       expect(envelopes).to.deep.eq(expectedEnvelopes)
     })
+
     it('emits messages for test run level hooks', () => {
       const envelopes = testEmitSupportCodeMessages({
         beforeTestRunHookDefinitions: [
@@ -278,6 +291,7 @@ describe('helpers', () => {
             id: '0',
             line: 3,
             options: {},
+            order: 0,
             uri: 'features/support/run-hooks.js',
           }),
         ],
@@ -290,6 +304,7 @@ describe('helpers', () => {
             options: {
               name: 'special cleanup thing',
             },
+            order: 1,
             uri: 'features/support/run-hooks.js',
           }),
           new TestRunHookDefinition({
@@ -298,6 +313,7 @@ describe('helpers', () => {
             id: '2',
             line: 11,
             options: {},
+            order: 2,
             uri: 'features/support/run-hooks.js',
           }),
         ],