+++ src/configuration/types.ts
@@ -64,6 +64,26 @@ export interface IConfiguration {
    */
   formatOptions: JsonObject
   /**
+   * Name/path of each plugin to use
+   *
+   * @example
+   * [
+   *   "\@cucumber/my-plugin",
+   *   "./custom-plugin.js"
+   * ]
+   * @default []
+   * @remarks
+   * Each item is a module specifier for a plugin to be loaded.
+   */
+  plugin: string[]
+  /**
+   * Options to be provided to plugins
+   * @default \{\}
+   * @remarks
+   * The value must be a JSON-serializable object.
+   */
+  pluginOptions: JsonObject
+  /**
    * Paths to where your support code is
    * @default []
    * @see {@link https://github.com/cucumber/cucumber-js/blob/main/docs/configuration.md#finding-your-code}