+++ CHANGELOG.md
@@ -8,6 +8,45 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 Please see [CONTRIBUTING.md](./CONTRIBUTING.md) on how to contribute to Cucumber.
 
 ## [Unreleased]
+### Changed
+- BREAKING CHANGE: Set `FORCE_COLOR` based on deprecated format option ([#2769](https://github.com/cucumber/cucumber-js/pull/2769))
+- Redesigned output for summary, progress, progress bar and pretty formatters ([#2708](https://github.com/cucumber/cucumber-js/pull/2708))
+
+### Deprecated
+- Deprecate legacy `SummaryFormatter` and `ProgressFormatter` classes in favour of new formatter architecture ([#2708](https://github.com/cucumber/cucumber-js/pull/2708))
+- Deprecate `printAttachments` format option in favour of `includeAttachments` ([#2708](https://github.com/cucumber/cucumber-js/pull/2708))
+
+### Fixed
+- Handle thrown strings correctly ([#2708](https://github.com/cucumber/cucumber-js/pull/2708))
+
+## [12.8.2] - 2026-04-25
+### Changed
+- Various dependency updates
+
+## [12.8.1] - 2026-04-14
+### Fixed
+- Resolve dependency issue with `junit-xml-formatter` and `query` ([junit-xml-formatter#143](https://github.com/cucumber/junit-xml-formatter/pull/143))
+
+## [12.8.0] - 2026-04-12
+### Added
+- Support custom externalising option ([#2780](https://github.com/cucumber/cucumber-js/pull/2780))
+
+## [12.7.0] - 2026-02-25
+### Changed
+- Warn about merged paths from config and CLI ([#2759](https://github.com/cucumber/cucumber-js/pull/2759))
+
+### Fixed
+- Handle ESM when building source references ([#2740](https://github.com/cucumber/cucumber-js/pull/2740))
+- Fix error type resolution ([#2744](https://github.com/cucumber/cucumber-js/pull/2744))
+- Ensure that environment variables reach child processes in parallel mode ([#2752](https://github.com/cucumber/cucumber-js/pull/2752))
+
+## [12.6.0] - 2026-01-22
+### Deprecated
+- Deprecate colorsEnabled format option in favour of environment variables ([#2729](https://github.com/cucumber/cucumber-js/pull/2729))
+
+## [12.5.0] - 2025-12-26
+### Added
+- Allow external plugins (see [documentation](./docs/plugins.md)) ([#2715](https://github.com/cucumber/cucumber-js/pull/2715))
 
 ## [12.4.0] - 2025-12-13
 ### Added
@@ -1710,7 +1749,13 @@ this.Given(), this.When(), this.Then() and this.defineStep() ([#2](https://githu
 
 ## 0.0.1
 
-[Unreleased]: https://github.com/cucumber/cucumber-js/compare/v12.4.0...HEAD
+[Unreleased]: https://github.com/cucumber/cucumber-js/compare/v12.8.2...HEAD
+[12.8.2]: https://github.com/cucumber/cucumber-js/compare/v12.8.1...v12.8.2
+[12.8.1]: https://github.com/cucumber/cucumber-js/compare/v12.8.0...v12.8.1
+[12.8.0]: https://github.com/cucumber/cucumber-js/compare/v12.7.0...v12.8.0
+[12.7.0]: https://github.com/cucumber/cucumber-js/compare/v12.6.0...v12.7.0
+[12.6.0]: https://github.com/cucumber/cucumber-js/compare/v12.5.0...v12.6.0
+[12.5.0]: https://github.com/cucumber/cucumber-js/compare/v12.4.0...v12.5.0
 [12.4.0]: https://github.com/cucumber/cucumber-js/compare/v12.3.0...v12.4.0
 [12.3.0]: https://github.com/cucumber/cucumber-js/compare/v12.2.0...v12.3.0
 [12.2.0]: https://github.com/cucumber/cucumber-js/compare/v12.1.0...v12.2.0